update main page styles, fix join game event
