+++ apps/shelter-client/src/styles/Main.scss
@@ -4,51 +4,66 @@
   color: #ffffff;
   margin: 14vh 0 0 0;
   padding: 0;
-  
+  font-weight: 500;
+
   .all-rooms-container {
-      background-color: rgba(0, 0, 0, 0.5);
-      padding: 40px;
+    background-color: rgba(0, 0, 0, 0.5);
+    padding: 40px;
+    font-size: 14px;
 
-      h2 {
-        margin: 20px 0 0 0;
-      }
+    hr {
+      border: none;
+      height: 0.5px;
+      color: #333;
+      /* old IE */
+      background-color: #fff;
+      /* Modern Browsers */
+      color: #fff;
+      margin: 6px 0 10px 0;
 
-      hr {
-        border: none;
-        height: 0.5px;
-        color: #333;  /* old IE */
-        background-color: #fff;  /* Modern Browsers */
-        color: #fff;
-        margin: 6px 0 10px 0;
-        
-      }
+    }
   }
 }
 
-.create-input {
+
+.create-room-container {
   display: flex;
   width: 30vw;
   margin: 0 0 10px 0;
+  font-size: 18px;
 
-  input {
-    flex: 1;
-    padding: 8px;
-    border: 1px solid #ccc;
-    background-color: #000;
+  .explore-text {
+    width: 69%;
+    cursor: default;
+    padding: 8px 10px;
+    color: #ffffffd9;
+    border: none;
     border-radius: 4px;
-    margin-right: 6px;
-    color: #ffffff;
-    outline: none;
+    font-weight: 700;
+    margin: 0 10px 0 0;
+    border: #a0a0a0 0.5px solid;
   }
 
   button {
+    width: 30%;
+    font-size: 18px;
     padding: 8px 10px;
-    background-color: #ffffff;
+    background-color: #faaa63ca;
+    color: #3c1f07d9;
     border: none;
     border-radius: 4px;
     cursor: pointer;
+    border: #1f1f1f 0.5px solid;
+
+    transition: transform 0.3s ease;
+    font-weight: 700;
+
+    &:hover {
+      transform: scale(1.1);
+    }
   }
 }
+
 
 .rooms-list {
   .room-item {
@@ -57,14 +72,14 @@
     text-align: center;
     align-items: center;
     background-color: #FFFFFF;
-    color: #000;
+    color: #000000;
     cursor: pointer;
     border-radius: 4px;
     padding: 6px 12px;
     margin: 4px 0;
-    border-bottom: 1px solid #dadce1 ;
-    -webkit-box-shadow: 2px 3px 7px -4px rgba(0,0,0,0.52);
-    box-shadow: 2px 3px 7px -4px rgba(0,0,0,0.22);
+    border-bottom: 1px solid #dadce1;
+    -webkit-box-shadow: 2px 3px 7px -4px rgba(0, 0, 0, 0.52);
+    box-shadow: 2px 3px 7px -4px rgba(0, 0, 0, 0.22);
 
     &:hover {
       background-color: #e3e3e3;
